requirements.txt
absl-py==2.1.0
agents==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
blinker==1.7.0
boilerpy3==1.0.7
Brotli==1.1.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
cryptography==42.0.5
cssselect==1.2.0
curl_cffi==0.7.0b4
databind==4.4.0
databind.core==4.4.0
databind.json==4.4.0
dataclasses-json==0.6.4
datasets==2.19.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==5.0.4
django-haystack==3.2.1
docopt==0.6.2
docspec==2.2.1
docspec-python==2.2.1
docstring-parser==0.15
duckduckgo_search==5.3.0b4
Events==0.5
exceptiongroup==1.2.0
executing==2.0.1
fake-useragent==1.5.1
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
flatbuffers==24.3.25
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-health-checking==1.62.1
grpcio-tools==1.62.1
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h2==4.1.0
h5py==3.11.0
haystack-ai==2.0.1
haystack-bm25==1.0.2
haystack-pydoc-tools==0.2.0
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.1.0
inflect==7.2.1
instructor==0.4.8
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.24
json_repair==0.11.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
jupyterlab_widgets==3.0.10
keras==3.3.3
langchain==0.1.20
langchain-core==0.1.40
langchain-text-splitters==0.0.1
langsmith==0.1.40
lazy-imports==0.3.1
libclang==18.1.1
llama-index==0.10.33
llama-index-agent-openai==0.2.3
llama-index-cli==0.1.12
llama-index-core==0.10.33
llama-index-embeddings-openai==0.1.9
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.16
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.19
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.8
llama-parse==0.4.0
llamaindex-py-client==0.1.19
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
nr-date==2.1.0
nr-stream==1.1.5
nr.util==0.8.12
num2words==0.5.13
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.28.1
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
playwright==1.42.0
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
prompthub-py==4.0.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.9.0b1
pydoc-markdown==4.8.2
pyee==11.0.1
Pygments==2.17.2
PyMySQL==1.1.0
pypdf==4.2.0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
quantulum3==0.9.1
ragas==0.1.7
rank-bm25==0.2.2
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
requests-file==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.19.0
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.5.0
sseclient-py==1.8.0
stack-data==0.6.3
streamlit==1.33.0
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
tomli_w==1.0.0
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
triton==2.3.0
typeapi==2.2.1
typeguard==4.2.1
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
wcwidth==0.2.13
weaviate-client==4.6.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
